# ABC-79: Optimize channel autocomplete query