#### Summary
As discussed in [this very long thread](https://pre-release.mattermost.com/core/pl/3nzompitxpfi8do54r1xtb7ooa), this changes the autocomplete query to return an arbitrary subset of results rather than the first X in the complete, sorted result set.

The impact on some autocomplete searches on my local database with over ~400,000 in-team channels prefixed with "channel":

| Search Term | Before | After |
|-|-|-|
| ~ | 7.65s | 0.021s |
| ~c | 9.48s | 2.83s |
| ~ch | 9.34s | 1.89s |
| ~cha | 10.60s | 1.48s |
| ~t | 7.68s | 0.592s |
| ~to | 8.32s | 0.027s |
| ~tow | 7.37s | 0.023s |

There are more things I could do to reduce those "~c..." cases down to a few milliseconds as well, but I'm suspecting that 400,000+ matching channels is not the norm, so it would probably be premature to put much effort into that.

#### Ticket Link
https://mattermost.atlassian.net/browse/ABC-79

I'm going to close this ticket with this PR as I think this resolves the biggest issues (The user performance isn't anywhere near as bad as the channel performance was.). We can create new issues as needed.

#### Checklist
N/A