@jasonblais 2018-01-29
@ccbrown Are we going to apply the same to user autocomplete? (and perhaps emoji autocomplete)