@ccbrown 2018-01-29
@jasonblais I think we should, but I think we should make separate JIRA issues and PRs for them.