@mattermod 2018-01-29
Spinmint test server created at: http://i-0f226cd0befe35620.spinmint.com

Test Account 1: Email: `test@test.com` | Password: `passwd`

Test Account 2: Email: `test2@test.com` | Password: `passwd`

Instance ID: i-0f226cd0befe35620