@jwilander 2018-01-30
@jasonblais as is, this is going to change more than just the behaviour of the channel autocomplete but also anywhere we have channel search as well. The case I'm thinking of is in the 'More Channels' modal, it's going to be much faster but it's going to show arbitrary results like it will for the autocomplete. Is that OK?

Personally I think it makes more sense to have a dedicated autocomplete API endpoint for this fast query and the regular search endpoint is slower but more accurate. Or at least it should be a query parameter on the search endpoint