@jasonblais 2018-01-30
@ccbrown Propose we only change the channel linking autocomplete. Other searches such as the "More Channels" dialog, results aren't expected to be arbitrary. Plus we might add paging to search results later, in which case we display more than just the first 100 results.

/cc @jwilander