@csduarte 2018-01-30
Is it possible the new behavior is going to be hard on users? Is this something that could/should be feature flagged? @ccbrown