@ccbrown 2018-01-30
@jasonblais This review now changes no behavior, but adds an additional API endpoint for channel autocomplete. I'll put in the reviews for the 3 other repos that will need to be updated soon.

@csduarte I don't really think anyone will notice a difference in the results; only that they'll get autocomplete results potentially several hundred times faster.