@mattermod 2018-01-30
Spinmint test server destroyed