@mattermod 2018-01-30
`Setup Test Server` label detected. Spinmint test server created if build succeeds (checks pass and no conflicts with base branch).