@jasonblais 2018-01-30
@ccbrown 

> This review now changes no behavior, but adds an additional API endpoint for channel autocomplete

Do you mean "More Channels" search still shows X arbitrary results? Or only the channel autocomplete.