@ccbrown 2018-01-30
@jasonblais I mean absolutely no behavior is changing with this PR. *Everything* is exactly as it was before. The behavior will change once I get the other PRs in and have the client use the new endpoint that this adds.