@jasonblais 2018-01-30
@jwilander @grundleborg Ready for your reviews.