@mattermod 2018-01-30
Spinmint test server created at: http://i-069cfc0b0ed2a6b37.spinmint.com

Test Account 1: Email: `test@test.com` | Password: `passwd`

Test Account 2: Email: `test2@test.com` | Password: `passwd`

Instance ID: i-069cfc0b0ed2a6b37